Improve naming of message tasks related to management clients filtering
